+++ src/groupeddataframe/groupeddataframe.jl
@@ -963,8 +963,13 @@ end
     filter(fun, gdf::GroupedDataFrame)
     filter(cols => fun, gdf::GroupedDataFrame)
 
-Return a new `GroupedDataFrame` containing only groups for which `fun`
-returns `true`.
+Return a new `GroupedDataFrame` containing only groups for which `fun` returns
+`true`.
+
+The use of this function is discouraged as it has API that is inconsistent with
+other transformation functions like [`select`](@ref). Instead it is recommended
+to use the [`subset`](@ref) function. The [`filter`](@ref) function is supported
+to ensure backward compatibility.
 
 If `cols` is not specified then the predicate `fun` is called with a
 `SubDataFrame` for each group.